Сode divided into the best file architecture
